requirements.txt
branca==0.5.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
folium==0.12.1.post1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
psycopg2-binary==2.9.3
requests==2.28.1
tqdm==4.64.1
urllib3==1.26.12
Werkzeug==2.2.2
